Handle cross-volume renames
